ibquery: completely rework command
This command now works like this:

  ibquery DEVICE [OPTIONS] [cmd...]

If no cmds are given after the end of options, there is a readline
prompt for commands to be entered interactively.

The former hyphenated commands are now bare words
  read/?, write, query, clear, delay, status

Further, unrecognized words are assumed to be queries if they end
in a question mark, otherwise writes.

So now this works

  ibquery gen id?

If no DEVICE is specified, list available devices.